test_rank and test_rank_regression_from so moved to test_reductions
