+++ sympy/matrices/tests/test_matrices.py
@@ -6,7 +6,7 @@ from sympy import (
     sqrt, symbols, sympify, trigsimp, tan, sstr, diff, Function, expand)
 from sympy.matrices.matrices import (ShapeError, MatrixError,
     NonSquareMatrixError, DeferredVector, _find_reasonable_pivot_naive,
-    _simplify)
+    _simplify, MatrixReductions)
 from sympy.matrices import (
     GramSchmidt, ImmutableMatrix, ImmutableSparseMatrix, Matrix,
     SparseMatrix, casoratian, diag, eye, hessian,
@@ -2615,35 +2615,6 @@ def test_simplify_immutable():
     assert simplify(ImmutableMatrix([[sin(x)**2 + cos(x)**2]])) == \
                     ImmutableMatrix([[1]])
 
-def test_rank():
-    from sympy.abc import x
-    m = Matrix([[1, 2], [x, 1 - 1/x]])
-    assert m.rank() == 2
-    n = Matrix(3, 3, range(1, 10))
-    assert n.rank() == 2
-    p = zeros(3)
-    assert p.rank() == 0
-
-def test_rank_regression_from_so():
-    # see:
-    # https://stackoverflow.com/questions/19072700/why-does-sympy-give-me-the-wrong-answer-when-i-row-reduce-a-symbolic-matrix
-
-    nu, lamb = symbols('nu, lambda')
-    A = Matrix([[-3*nu,         1,                  0,  0],
-                [ 3*nu, -2*nu - 1,                  2,  0],
-                [    0,      2*nu, (-1*nu) - lamb - 2,  3],
-                [    0,         0,          nu + lamb, -3]])
-    expected_reduced = Matrix([[1, 0, 0, 1/(nu**2*(-lamb - nu))],
-                               [0, 1, 0,    3/(nu*(-lamb - nu))],
-                               [0, 0, 1,         3/(-lamb - nu)],
-                               [0, 0, 0,                      0]])
-    expected_pivots = (0, 1, 2)
-
-    reduced, pivots = A.rref()
-
-    assert simplify(expected_reduced - reduced) == zeros(*A.shape)
-    assert pivots == expected_pivots
-
 def test_replace():
     from sympy import symbols, Function, Matrix
     F, G = symbols('F, G', cls=Function)
@@ -2710,7 +2681,6 @@ def test_pinv():
         comp(i.n(), j.n())
         for i, j in zip(q.subs(reps), w.subs(reps))
         )
-
 
 @XFAIL
 def test_pinv_rank_deficient_when_diagonalization_fails():
